Fix partial discards issue on mvme3100
Seems to be something strange going on with the network stack just after
iocInit. Getting some garbage data (maybe duplicate packets?) leading
to discards. The network stack seems to hold onto the data for a while
leading to EAGAIN. Retrying seems to resolve the problem.

RPCIO driver for NFS v2 also shows a similar error with an unexpected
transaction ID.