t9p: Traverse symlinks manually in t9p_open_handle with Twalk
9P2000.L does not transparently handle symlinks/links like I thought.
Trying to descend into a symlinked directory using Twalk will return
nwqid < nwcomps, where qids[nwqid-1] is a symlink or hard link.
Previously that was treated as an error. To fix this, we'll evaluate
the symlink, splice it into the path, and walk again until we end up
at the final path.

This isn't so different from normal path evaluate in RTEMS. Maybe
a refactor could be made there...

WIP fixing symlink traversal